DifferentialDiagnosis/DifferentialDiagnosis-Part2/requirements.txt
google-cloud-storage==1.37.1
numpy==1.19.5
pandas==1.2.4